backend/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "face-recognition-attendance"
version = "1.0.0"
description = "Face Recognition Attendance System"
readme = "README.md"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
license = {text = "MIT"}
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn>=0.23.2",
    "python-multipart>=0.0.6",
    "face-recognition>=1.3.0",
    "opencv-python>=4.8.1.78",
    "numpy>=1.26.0",
    "pydantic>=2.4.2",
    "PyYAML>=6.0.1",
    "python-jose>=3.3.0",
    "passlib>=1.7.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.3.0",
    "isort>=5.12.0",
]

[tool.setuptools]
packages = ["face_recognition_attendance"]

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 88
